add fitur download room design
